Update functions and change hp and base_dmg type

+++ dungeon_despair/domain/entities/enemy.py
@@ -8,10 +8,10 @@ from dungeon_despair.domain.entities.entity import Entity
 
 class Enemy(Entity):
 	species: str = Field(..., description="The enemy species.", required=True)
-	hp: int = Field(..., description="The enemy HP.", required=True)
+	hp: float = Field(..., description="The enemy HP.", required=True)
 	dodge: float = Field(..., description="The enemy dodge stat.", required=True)
 	prot: float = Field(..., description="The enemy prot stat.", required=True)
-	spd: int = Field(..., description="The enemy spd stat.", required=True)
+	spd: float = Field(..., description="The enemy spd stat.", required=True)
 	attacks: List[Attack] = Field([], description='The enemy attacks', required=True)
 
 	def __str__(self):
